feat: Update to 1.21.5
Updated to Fabric and NeoForge for 1.21.5

Removed item properties; now, we use the new item model format to
check for excited slimes.

Reworked item registration for the new system.

Slime chunk component is now either present or removed, instead of
true/false.

Removed obsolete accessors.

Use a getName() override instead of appendTooltip hack for named
bucketed slimes.

Trails now properly come from the players hand, and show up for
other players. This took a lot of horrible mixins - not sure how
this is for perf and compat. Seems to work with Sodium though.
In first person, we do some calculations with the camera vectors to
position the particles. In third person, and for other players, we
cache the transform of the PoseStack used to render items in player's
hands to get the exact position of the hand, and then use that
position to spawn our particles. This requires storing a reference
to the current player in PlayerRenderer, and storing the last off and
main hand positions in AbstractClientPlayer.